requirements.txt
grpcio==1.44.0
grpcio-tools==1.44.0
mysql-connector-python==8.0.27
